requirements.txt
asgiref==3.2.7
beautifulsoup4==4.9.1
certifi==2020.4.5.2
chardet==3.0.4
Django==2.2.13
django-bootstrap-pagination==1.7.1
django-bootstrap4==2.0.1
django-cors-headers==3.7.0
django-extensions==2.2.9
django-webpack-loader==1.1.0
djangorestframework==3.12.4
Faker==4.1.0
idna==2.9
importlib-metadata==1.6.1
mysqlclient==2.0.3
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.9
zipp==3.1.0
